intermediate changes ref:8f1c383d5266ee01c50049be4ceb539c78297914

+++ library/cpp/testing/unittest/registar.h
@@ -410,13 +410,13 @@ public:                       \
         const auto _ed = (E);                                                                                                          \
         const auto _ad = (A);                                                                                                          \
         const auto _dd = (D);                                                                                                          \
-        if (std::isnan(_ed) && !std::isnan(_ad)) {                                                                                     \
-            const auto _as = ToString((long double)_ad);                                                                                            \
+        if (std::isnan((long double)_ed) && !std::isnan((long double)_ad)) {                                                           \
+            const auto _as = ToString((long double)_ad);                                                                               \
             auto&& failMsg = Sprintf("expected NaN, got %s %s", _as.data(), (TStringBuilder() << C).data());                           \
             UNIT_FAIL_IMPL("assertion failure", failMsg);                                                                              \
         }                                                                                                                              \
-        if (!std::isnan(_ed) && std::isnan(_ad)) {                                                                                     \
-            const auto _es = ToString((long double)_ed);                                                                                            \
+        if (!std::isnan((long double)_ed) && std::isnan((long double)_ad)) {                                                           \
+            const auto _es = ToString((long double)_ed);                                                                               \
             auto&& failMsg = Sprintf("expected %s, got NaN %s", _es.data(), (TStringBuilder() << C).data());                           \
             UNIT_FAIL_IMPL("assertion failure", failMsg);                                                                              \
         }                                                                                                                              \
